scripts/sports_center/requirements.txt
ae-ffmpeg==1.1.2
auto-editor==23.51.1
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2023.12.2
greenlet==3.0.3
idna==3.6
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
mutagen==1.47.0
mysql-connector-python==8.4.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
pandas==2.2.2
Pillow==10.1.0
pyav==12.0.0
pycryptodomex==3.19.1
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.30
sympy==1.12
tiktoken==0.5.2
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
websockets==12.0
yt-dlp==2023.12.30
